Consolidate GPGPU apt filtering functionality into detect.py
Prior to this change, the gpgpu driver installation procedure
involves additional filtering of package candidates in
./ubuntu-drivers, which the test cases (which are testing for
 end-user package filtering expectations) and any other direct
callers of detect.py's filter functions are unaware of.

This change moves all filtering logic to detect.py and adjusts
all associated test cases as needed to represent the actual view
of which drivers the end-user will have installed.

This change also moves the runtimepm check outside of the package
loop in get_desktop_package_list